requirements.txt
# Database and ORM
sqlite-utils
sqlalchemy
libsql-client
psycopg
xata

# API Requests and Data Handling
requests
pandas
pybaseball
MLB-StatsAPI
python-mlb-statsapi
datetime
selenium
webdriver-manager
html5lib
cloudscraper
curl_cffi
more_itertools
timezonefinder
tqdm

# Environment Variables
python-dotenv

# Jupyter Notebook
jupyter

# Model
scikit-learn
xgboost
joblib
seaborn
matplotlib
